Global Automotive Drive Axle Market Research Report 2024-2030
Automotive drive axle is an integral component of vehicles, especially for the rear wheels. Its main function is to transmit driving torque to the wheel, as well as to maintain the position of the wheels relative and the vehicle body. Automotive drive axle is mainly used in the commercial vehicles. While for the passenger vehicles, such as multi-purpose vehicles (MPV) and minibus, automotive drive axle is also a necessary component.